od6.0: OD-1488, Volume Builder: minor UI issue in WellLog Interpolator
Summary: A minor update issue was present in well-log interpolator. On setting algorithm to anything except Inverse distance. While selecting the algorithm the radius field was was set to hide, but while reloading the data, the search radius field was set to inactive but was displayed.

Test Plan: The display is as desired. If the algorithm selected is not Inverse Distance then the field wont be displayed

Reviewers: ranojay

Reviewed By: ranojay

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11322