[od6.0] Thalweg horizon did not autoshow horizon data items (reopening OD-760)
Summary:
Thalweg horizon did not autoshow horizon data items (reopening OD-760).
Caused by previous fix r44637/D10517 that prevented all items of a horizon
in the data tree to be copied to a newly made tree item of that same horizon.
Solved by having  Thalweg plugin set the new AutoShow type as its AuxDataType.

Test Plan:
Tested by having the Thalweg plugin create a horizon. It now comes with a
whole list of horizon data items. Tested that normal horizon data items
are not copied to a newly made tree item of that same horizon, proving
that the previous fix has not been undone. Also tested that horizon data
does not autoshow in case a color blended version of that same horizon is
added to the tree. Only tested thoroughly in OD6.0. The trunk shows a lot
of strange behaviour in the data tree, but that turns out the same without
this new revision.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11325