[od6.0] Interactive dragging speedup with faults in scene (part of OD-660)
Summary:
Interactive dragging speedup with faults in scene (part of OD-660).
(1) Added isOn() checks to skip intersection line calculation entirely in case
faults are disabled in tree while display only-at-sections is selected.
(same as for 3D horizons in Phabricator D10937).
(2) Added areSticksDisplayed() checks to skip determination of intersecting
fault sticks and knots in case display of stick is disabled in tree.

Test Plan:
Tested with F3_Demo-CSM, which has a large set of faults
(1) Displaying faults at sections and next disabling them in the tree
results in smoother/faster dragging of inlines with preloaded data. Next
re-enabling them in the tree displays them at the correct locations, which
proves that an update is triggered in time anyway.
(2) Now smoother/faster dragging of inlines with preloaded data when faults
are displayed without their sticks. This both holds if faults are displayed
in full or only at sections. Next enabling stick display in the tree shows
the sticks and knots at the correct locations, which proves that an update
is triggered in time anyway.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D11180