od6.0: Solved crash in saving an horizon, or re-starting the tracking
Summary: the tracking TrcKeySampling does not have to be identical to the horizon geometry TrcKeySampling. Hence it is required to check if a node is included in the tracking sampling before reading one of the 3D arrays whose size matches the number of elements of the tracking TrcKeySampling. Fixes OD-1493.

Test Plan: no more crash with the option "Save As" or "Start tracking".

Reviewers: ding

Reviewed By: ding

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11349