od6.0: Fixed removal of output datapack in VolumeProcessing attribute application
Summary: the datapack was referenced by the chain executor, a local object, not by the chain, a class member. Last week I fixed the missing removal of the output reference in the chain executor destructor, which is why I need to adapt this part now so the output datapack does not get deleted.

Test Plan: applied a volume builder setup on an inline of F3_Demo. Could see the output, and no PE message when closing OD.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D11362