od6.0: Fixed application of wellLog interpolator with existing wells loaded
Summary: the interpolation initialization returned false if wells were loaded in the manager because their data was not used to initialize the log and track of WellLogInfo class. Also the return false error was not caught. First step for solving OD-1483

Test Plan: can compute the setup in a survey with/without wells in the tree.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D11363