[OD6.0]OD-1363 Crash during picks edit using lasso select tool
Summary:
The bug I can re-produce it in two ways:

    add an existing horizon, then add a horizon aux data item, and then using polygon tool

to remove part of horizon.

    add an existing horizon, and start to tracking, and then using polygon tool to remove horizon area.

At the end, I found the problem which caused by a negative index from the function
nearestIndex(). Form example: rrg.nearestIndex( bid.inl() ), rrg is (257,539 ) with step 1, bid is (256,539), the return value is -1.

The solution can fix this bug for the time being. But eventually, I think the problem is caused
by an offset from the function getBinID() in BinIdValueSet. and that is could be caused by how is the data added in the BinIdValueSet.

Test Plan: Tested in F3-demo in both of above two ways.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11353