od6.2: Show number of components in display name
Summary:
Show number of components in display name and which one is currently displayed. So when 5 components are loaded and the 2nd one is displayed, the display name starts with '2/5'.
Fixes OD-1338

Test Plan: Loaded several frequencies from a specdecomp volume. Item shows correct number of components and the current one, which changes when pressing page-up/down.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11424