[OD6.2]OD-1492 3D Horizon Auto-tracking: issues when resume tracking a horizon with empty QC grids
Summary:
This is ported from the revision D11378.
Based on the description from the Jira issue, the horizon doesn't have auxdata because it didn't do auto tracking.

when doing re-enable track after re-load the horizon,

    we need add the default auxdata in the tree that means the auxdata can not be removed if the reader can't read auxdata

so I added an extra function setEnableRemoveAuxData() in class surfaceAuxData.

    the current executorgroup class will be failed once there is one executor is failed. In some case we don't want to this result.

so, in this revision I added an extra function: setSkipSingleExecutorError.

The changes should not have any influence on existing OD features.

Test Plan: Tests following the steps from the Jira issue. It works fine.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11439