od: Added Seis::ProviderSet to get data from multiple providers through various modes.
Summary: At present, works for 3D in two of the modes with a minor problem. Have to enable it for other modes and for 2D in further revisions.

Test Plan: Tested by merging 3D seismic data. At present, skips the last one in the merged data. Looking for where the problem lies.

Reviewers: bert

Reviewed By: bert

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11457