[od] Unwanted data loading while scrolling Z-slice through preloaded data cube that was stored on disk with non-unit z-step (OD-782).
Summary:
Unwanted data loading while scrolling Z-slice through preloaded data cube
that was stored on disk with non-unit z-step. Should just behave the same as
when data cube is stored on disk with unit z-step, but then preloaded with
non-unit z-step subselection. So no pop-ups of data loading dialogs and display
of undefined colored intermediate slices, but round-off to the closest z-value
for which data is available. One might prefer a kind of interpolation in the
future if that can be implemented efficiently enough not hurt a real-time
dragging experience.

Test Plan:
Tested in survey OD-782 with 4m z-step. Added z-slice with as attribute the 
fully pre-loaded 21 VCF Thr7850 Size 200 BodySize cube with 8m z-step. When
dragging it through the scene, no more pop-ups of data loading dialogs and
no more displaying of undefined colored intermediate slices.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11475