od6.0: Fixed 'Not saved' error
Summary: Fixed 'Not saved' error as described in OD-1262. Also prepared for adding timegate to pseudo relief attribute (OD-1448) which needs a change in order of some calls in createAttribDesc

Test Plan: Tested the 'Not saved' bug. Opened a stored set. Name stays in the name box, regardless of any changes made.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D11498