od6.0: Fixed the velocity gridding using RBF gridder
Summary: the gridder was used with the wrong assumption that the weights were always independant of the input set of values. As that is not the case for the RBF gridder, fix it. Fixes OD-1100.

Test Plan: applied RBF gridder on inline 425. Results are similar to the triangulation as expect when using a regular set of functions.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11548