od: improved user-friendliness of the Progress Viewer window
Summary: viewer status bar now says if the process is running, was terminated or finished successfully. Greatly improved the display speed by switching from uiTextEdit to uiTextBrowser. Partly corrected the issues with not being able to scroll up the slider. Fixes OD-1201.

Test Plan: all working except for the connection with the Qt signals. This prevent scrolling up or down for a longer period than set in the timer (0.5 sec).

Reviewers: bert, nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11576