od: OD-1106, In depth scene mouse reader on the well shows incorrect MD values in the status bar of OD Main
Summary: There was a loss of precision because of the large difference between two dah points data. While viewing in time scene the data was resampled and the precision of data increased. Now 3d geomtery is being used to find the accurate point

Test Plan: For straight wells it is producing results same in all the scenes Also created a curved well track and md values were seen changing as expected

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11597