od: Further improvements in od_ProgressViewer
Summary:
one could no longer copy/paste text especially during update of the content. Fixed that. Also added some help if the application is run without or with improper arguments.
Finally added the possibilities to refresh the process ID by monitoring the log file after sucessfull finish, and to kill the process without closing the progress viewer.

Test Plan: works as listed above. started a process until the end. Restarted it from the terminal while the viewer was still open, terminated it from the viewer. All as expected. A slider/mouse action resumes the update of the viewer after a selection was made.

Reviewers: nanne, bert

Reviewed By: bert

Subscribers: bert

Differential Revision: https://phabricator.dgbes.com:8443/D11610