[OD] re-work about undo issue in em manager
Summary:
As we discussed, the previous undo mechanism is not suit for multi em objects. The idea of the new work is trying to keep each 
emobject has one undo. later on, I will add the max length setting in undo when it is horizon 3d em object.

Test Plan: Tests with other changes ( coming revsions ). it works fine

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11571