[OD6.2] OD-1554 Polygon Selection / Remove tool: removed part is different from initial selection if camera position changed after selection and before removal.
Summary:
In the solution of OD-1527, we removed camera clone function because it takes more time when the scene contains many objects or big objects. so a new selector still contains the old camera's position.
It uses the old position to check if a point in the selected polygon that can not be corrected any more after user changing camera position.

The previous design we keep every selector and remove selections using these selectors. But since we implemented displaying selections I think we don't need to keep all selectors any more. we can use
selections to instead of selectors.

comments: the old function: removeSelected(...) in emobject and emhorizon2d should not be used any more.

Test Plan: Test in F3-demo in both of 3d and 2d horizon. it works fine.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11546