od6.2: Bulk fault export
Summary:
The export fault and faultstickset did not have any option for export of faults in bulk. This implementation lets us export multiple faults and faultstickset in one go. Although implementation of faultstickset is implemented but since the no import facility is present it is not shown in the menu

Final commit bulk work

Test Plan: The faults are being exported and proper *.dat file is being created for both single export and bulk export. On import the faults are visible properly. tested in F3Demo2012 FaultA

Reviewers: raman

Reviewed By: raman

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11633