od6.2: Made attributeset manager aware of 2D and 3D
Summary:
Main change in this revision is that I made a separation between 2D and 3D attributesets in its manager. For this I also changed the Survey - Manage item to have a 2D and3D submenu. In the manager I made an improvement to the info text.
PS: As I am preparing for a 2D depth scene, this change came as well in this revision.

Test Plan: Checked Survey-Manage menu and checked if the 2D and 3D windows show different information. Checked the info field for its stored data information, which is good now.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D11659