od: Added verbosity for the RadialBasisFunction gridder de-trending interface
Summary: the RBF de-trending cannot always be done if the number of input points is too small. Result is that a high order could silently be reverted to a simple average removal. Changed the interface to make the user aware of this restriction. Fixes OD-1487

Test Plan: tested from well-log interpolator and basemap interfaces. Got the expected warning (when required)

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D11688