[OD]OD-787 further fixing
Summary:
We have done revision D11653 which fixed the Jira issue OD-787. But it creates another bug which makes all other survey object can not be removed
from tree item. That is because the return from the function canRemoveAttrib() perhaps always is false due to no attribs. 

So the new fixing is add a new virusal function in visSurvey object named as canBeRemoved()... Then we use canRemove(display ID) instead of canRemoveAttrib in uiDispalyTreeItem and uioddatatreeitem.

Test Plan: Tested working fine

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11694