od6.2: Removed unnecessary offset ft to m conversion
Summary:
For prestack data, the offset was always converted to meter. This is unnecessary and leads to conversion errors. It will now only convert if the survey XY unit is different from the offset file unit.
I think the new implementation is correct and more in line with our unit policy. Would we need backward compatibilty anywhere?

Test Plan:
In my test survey, which has XY in feet, the mousepos info is correct again in 2D and 3D view. The component names are also right now.
Tested the prestack attribute with an offset range. Works as expected.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D11698