[OD6.2]OD-787,OD-630 and OD631: further more
Summary:
This revision is corrected the errors based on the rule of keeping ABI.
After talking with Raman, we can add argument in virtual function in od6.2. so I added 
new functions in this revision, and also did related changes due to the new function.
Another things, in this revision, I recovered the removed part which was done by Nanne 
for fixing the compile error.

Test Plan:
Tested by debug I can see the three node source type are setted in different interpolation: manual, 
auto and gridding

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11711