od6.2: Fixed wavelet stacking
Summary:
This is quite a mess. Cleaned-up a bit, but a complete rework in the trunk is needed.
Things went wrong with gui selections vs checks, wavelet size, wavelet sampling and drawing of the stacked wavelet.
Stacked wavelet is now always available, not only when 2 or more wavelets are checked. This made management easier.
Related to OD-1620

Test Plan: Stacked wavelet is visible again. Correct sampling is used and it's saved correctly.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D11779