od6.2: Added a function to stop the pre-loading of the data when pressing abort
Summary: the actual work is forwarded to the threadwork manager. Hence if the execution is not sucessfull one must empty the queue because a number of tasks may still be running. Otherwise the desctructor will wait for all to be finished. Done by re-implementing the goImpl function.

Test Plan: tested by pressing abort in interface, software more responsive and all tasks get stopped (frees CPU and disk usage, checked with monitors).

Reviewers: helene

Reviewed By: helene

Subscribers: mahant

Differential Revision: https://phabricator.dgbes.com:8443/D11866