od6.2: Fixed the pre-loading of 2D data for Volume processing
Summary: the forecast of the trace positions to be read was too simple. Now it replies on information provided by the translators of each line.

Test Plan: tested BG SteeringCube creation on 10 lines, some of which having decreasing trace numbers, gaps in the trace numbers (several sections per line), and several traces with identical positions (all three on the same line!). There are still 5 traces which output a pErrMsg, but maybe the information in the CBVS header is wrong since this is very old data.

Reviewers: bert

Reviewed By: bert

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11921