od6.0: Fix for horizon intersection displays in 2D viewer
Summary: If horizon is only available at 1 position, it should be displayed as a marker. Fixes OD-1580. Need to check if it works in 6.2 and what kind of fix is needed there.

Test Plan: Add inline and pick horizon there. Add crossline and open in 2D viewer. Intersection marker is visible now.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11940