od6.2: OD-1106, In depth scene mouse reader on the well shows incorrect MD values in the status bar of OD Main
Summary: There was a loss of precision because of the large difference between two dah points data. While viewing in time scene the data was resampled and the precision of data increased. A separate dah series is created and resampled.

Test Plan: The fix now shows depth in depth scene almost equal to depth in time scene. tested using F3Demo2012, F034

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11957