od6.2: Fixed differences between velocity gridder and well log interpolator
Summary: some differences remained in the way the functions are extrapolated or the layermodel is used. Now the results are fairly similar, with the exception that well logs are not very well sampled when vertically sparse (the LogSampler does not treat them as interval quantities). Improves the initial fix for OD-1481.

Test Plan: applied Horizon/HorizonCube based gridding with both modules.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D11979