od: Made the initialization of datapack values selectable
Summary: most datapack users do not need to have the values initialiazed. For the Seis::Loader implementation this work is send to the Thread::WorkManager at the start of the execution, while the stream is slowly being read. This eliminates all initialization time. Implements part 2 of OD-1775.

Test Plan: tested by pre-loading an irregular 25GB dataset from the survey menu and of a 2D dataset with several lines, and running a Dip-steering creation volume processing.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D11997