od6.2: remaining problems in inverse distance gridding
Summary: an epsilon value was too small, as the weights can quickly drop to 1e-17 according to the new equation with radius. Second issue solved: the weights were not correct in case of equidistant points, as the sum of all weight should be equal to 1.

Test Plan: removes the two artifacts seen by Ayrat when testing OD-1474

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D12037