od: OD-1724, Basemap: mouse pointer in regional 2D surveys is poorly visible
Summary: The cursor size, shape and color could not be updated according to the users wish but it could be done now. This also improves the visibility of the cursor when multiple data is viewed on the Basemap

Test Plan: Tested using MNSH_OGA_2016_v1 survey. Everything is working as per required.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12063