od6.0: Fix for corrupted Horizon Data after renaming [OD-1791]
Summary: While renaming a Horizon Data (attribute) we modify the ascii header of the file and then copy the rest of the data (mostly binary) from the original stream. For some reason, this 'copy' was modifying the content, leaving a corrupted file. This change to binary functions fixes the issue. In any case, I would rather go binary while making a blind copy.

Test Plan: Renamed several Horizon Data and the renamed attribute is still usable. Also confirmed by byte-by-byte comparison that the copy is now idential to the original.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D12066