od: Fixed erroneous error message on OpenGL version
Summary:
ui3dviewer incorrectly reports version<2, because the string "4.4.0" cannot be converted to a float by QString. There seems to be a discrepancy in implementation between:
string2.h: mGlobal(Basic) bool getFromString --> never uses the QString tools
convert.h: Conv::to<float>( str ) --> uses QString object if Qt is available.
I added a fallback to the non Qt implementation for this latter.

Test Plan: warning is gone. Note that on windows the warning dialog was blocking the application, and neither the warning dialog nor od_main could be accessed/closed.

Reviewers: nanne

Reviewed By: nanne

Subscribers: bert

Differential Revision: https://phabricator.dgbes.com:8443/D12136