od: Write relative path in sgydef file
Summary: Added call to SEGY::FileSpec::makePathsRelative to make sure the path to the sgy file is relative. This change only affects the file name written in the footer of the sgydef file. In the header we save the same filename. Seems we don't use it when reading the sgydef file?

Test Plan: Imported segy file in 'scan' mode. Data loads fine. Moved survey to another location. Data still loads fine.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D12149