od6.2: Finalizing shown MainWin
Summary:
a uiDialog which is constructed and thereafter only displayed with the "show" function never gets finalized before being shown, resulting in potential layout issues. This fix ensures the dialog will be shown finalized. The only other alternative would be to call "goMinimised" if the construction is not done at the same place as the call to "show".
Please approve or reject.

Test Plan: layout of my uiDialog is not correct without call to "goMinimized". The uiDialog is created in the constructor, and the call to "show" is in the implementation of the CallBacker function linked to a push button.

Reviewers: nanne, bert

Reviewed By: bert

Subscribers: raman, bert

Differential Revision: https://phabricator.dgbes.com:8443/D12150