od6.2: Listbox selection storage looks at 2D geometry now
Summary:
The listbox selection storage for 2D lines in the 3D scene was different from the one in the basemap which looks at 2D geometry. Fixed this. See for detailed explanantion OD-866 and OD-1817.
What should we do in terms of backward compatibility? Should we convert the old objsel files into the new format?

Test Plan: Object selections made in the 3d scene can be used in basemap and vv.

Reviewers: raman, arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12148