od6.2: Change markertype for large pointsets and disable switch to pick mode
Summary:
Pointsets larger than 1000 will automatically displayed as points for performance reasons.
At the same time fixed a long time complaint that selecting a pointset in the tree always changes the view mode to Interact mode. This often leads to unwanted extra points by accidentally clicking in the scene.
The change only happens now when the pointset is empty, eg when adding a new pointset.
In the future we should add an edit mode to pointsets, similar to the tracking mode for horizons.

Test Plan:
Created a set of 50000 points. Display in the tree is now fast and doesn't stall the system anymore.
Selecting another pickset does not change the viewmode.

Reviewers: raman, arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12159