od: Fix for slow histogram popup
Summary: 8 bit data has many duplicates which is a problem when sorting. For calculating statistics, applied the same fix as we did in the DataClipper. A duplicateSort is done when the data is 8bit. Fixes OD-1836

Test Plan: Tested on Maui-NZ_SEG2017 survey. Histogram window pops up after about 1s.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12169