od6.2: Fix for crash in T2D using Vrms
Summary: If you create a depth scene using Vrms and then add an inline with default data, you get a crash. The VarLenArr stuff creates an array which gets deleted after it is out of scope. In this case the array 'ptr' is used later on.

Test Plan: No more crash with Vrms depth scenes.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12173