od6.2: Fixed Z idx and initialization issues in SequentialReader
Summary: If the DataPack Z range is larger than the seismic cube, the Z index was not right. Also, in this case we need to initialize the Arrays with udf as the reader will not fill the whole array. This fixes the Z shift issues in many volproc steps (OD-1842)

Test Plan: Tested VoxelConn and Smoother outputs. They look good. Also tweaked the if condition to force the else block to execute and confirmed that no change is needed there.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12177