od6.0: Bugfix for rubberband zoom
Summary: When moving mouse over a 2D intersection line in a 2D viewer, the drag mode is set to NoDrag. If previously in RubberBandDrag mode, it did not reset correctly if you're also in editmode. It seems being in editmode is not relevant, so removed this check. Fixes OD-1833.

Test Plan: Start tracking a 2D horizon in a 2D viewer. Enter rubberband zoom mode and move mouse over 2D intersections. Zooming is still possible.

Reviewers: satyaki

Reviewed By: satyaki

Differential Revision: https://phabricator.dgbes.com:8443/D12193