od6.2: Caching LatLong proj for re-use
Summary: While converting coordinates, we usually need the LatLong variant of either the source or the target. Calling pj_latlong_from_proj has its own overhead if you are doing the conversion in a loop. After this change, we always have a LatLong variant of the projection (if applicable) as class member for re-use.

Test Plan: filling a PickSet from a Shapefile is 5 times faster now.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12245