od6.2: OD-1839, Transform of X/Y from/to Lat/Long in File mode needs improvement
Summary: Solved the problem with format of the file, generally accepted long lat format is now being used along with a message regarding the format of the file. There is no restriction on the number of the columns a file has.

Test Plan: Tested, proper file now being made and message being diplayed

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12244