od6.0: fixed wrongly set debug count variable Threads::Mutex class
Summary: the implementation of Threads::Mutex::tryLock did not increase the count, unlike in Threads::Mutex::lock. As a result there was a negative count while desctructing the object. and lockingthread_ was not set to 0 as it should be. Maybe the result for a number of error flagged by valgrind on this object.

Test Plan: found after I had inserted the forced crash as a result of valgrind hints.

Reviewers: bert

Differential Revision: https://phabricator.dgbes.com:8443/D12253