od6.2: Revise memory estimation method in VolProc::ChainExecutor
Summary: the previous method did not take into account the cascading effect of stepouts, and impacts of the connectivity between epochs, where not all datapacks of a previous are released.

Test Plan: tested TFL crash in B13 survey: The chunk size is reduced, 85% of the system memory is used during the execution of the most memory consuming task of the chain (which is not the one with most working arrays), less otherwise.

Reviewers: bert

Reviewed By: bert

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12277