od6.2: Lazy initialization of Proj4Projections
Summary: When CRS plugin is loaded we make a repository of all available projections. This is mysteriously slow on some Windows machines. So, now we create the repository with only definition strings and the actual Proj4 objects are created only when needed by a particular transform operation.

Test Plan: CRS plugin initialization is 4 times faster on my Linux machine. But the real test will be done with a Windows release. Tested converted Lat/Long and X/Y values in UI and found OK. Also test_crs still gives success.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12285