od6.2: Fixed removal of temporary output files
Summary: MMP saves chunks of data in a temporary folder. This folder was not deleted after a successful run. The name of the ioobj is not i.* but i.[first inline number]. Fixes OD-1723.

Test Plan: Processed an Energy attribute in MMP mode. Temporary folder is now deleted at the end.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12288