od6.2: Fixed 2D hor loading with missing lines [OD-1894]
Summary:
If the line geometry of one or more lines is missing, reader will skip the lines and we will still have a working horizon on the remaining lines.

Also changed text of the warning message while deleting 2D lines and added the word 'interpretation' there.

Test Plan: Deleted a line form F3, the 2D horizons still load on the remaining lines. Also created a HorizonCube successfully on the remaining lines.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12296