od6.2: Fix for writing depth converted seismics to segy directly
Summary: There's a usePar missing to give the translator the correct sampling. Fixes problem when writing a depth converted volume directly to SEG-Y as described in OD-1908.

Test Plan: Process time to depth volume and output as SEG-Y. Then import the same segy file, and checked the sampling to verify it's written correctly now.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D12301