od6.2: Added saveProcPars
Summary:
Made history recording easier by adding a function to uiBatchJobDispatcherSel. Will save a proc file with the same basename as the output file.
Re-creating attributesets from the proc files was already in place. Need to have it for volumebuilder chains as well.

Test Plan: proc file is saved (again) when processing attributes or volumebuilder chains.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12299