[od6.2] Bugifx for showing markers at start of polyline
Summary:
Reverted a change from Ding. In his code, the HorizonPainter3D created a marker at every new beginning of a polyline. I assume to make sure that single point intersections are visible. I removed this code and added some intelligence to the uiAuxDataDisplay. If a polyline has a size of 1, a marker is drawn.
JIRA: OD-1902

Test Plan: Displayed some 3D horizons to a 2D viewer. The markers at the beginning of a polyline are gone. Other test: Started picking a new horizon on an inline. Then displayed a crossline and showed in a 2D viewer. Intersection point is visible.

Reviewers: prajjaval

Reviewed By: prajjaval

Differential Revision: https://phabricator.dgbes.com:8443/D12303