[od6.2] Well display related fixes
Summary:
From OD-1943: Renamed tree item 'New Well Track' to 'Pick New Trajectory'. Added label to left of list with existing wells. Did not work on the first item of this issue, because I'm not sure this is what we want.
From OD-1920: Set MultiID to display object before adding it to the scene.

Test Plan: Did a gui check. Looks better now. With fix from basemap, the wells are added again when opening the basemap.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12313