od6.2: Fix for empty line names in Manage Hor 2D lines window [OD-1894]
Summary: We have to read the missing linenames/GeomIDs in the info section of the reader, or else the indexing would get messed up while reading the data. So, this was an easy to filter out the invalid lines while passing this info on to the user.

Test Plan: Tested by removing line geometries of a few lines in nlog survey, the empty lines are gone. Also, the Z values for the other lines are correct which confirms that the indexing is still OK.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12369