od: Various fixes in limitTo functions
Summary: Virtualized the limitTo function in ranges.h. Existing implementations in TrcKey(Z)Sampling classes were happily shifting the sample distributions, and/or changing the step, which is not the purpose of these functions. I have some fears that this fix will introduce few bugs because of bad usage, like in SurveyInfo::setRange where the limitTo function was used to initialize the TrcKeyZSampling, but let's figure out in due time. Fixes OD-1918 (together with another bug fix for OD-1972).

Test Plan: added some tests to the existing test program. Ran a VolProc PCA computation where previously a shift was introduced by the previous implementation. Loaded an inline, a 3D horizon in the scene.

Reviewers: bert

Reviewed By: bert

Subscribers: nanne, raman

Differential Revision: https://phabricator.dgbes.com:8443/D12373