[od6.2] Fix for 2D ZAxistransform of points
Summary: PointSets are not transformed in a Wheeler scene when HorizonCube is 2D. canTransformSurv is now telling me whether the transform is 2D or 3D. Then needed to loop through all 2D lines to find the nearest TrcKey. Is this the most efficient way?

Test Plan: PointSets displayed correctly in 2D wheeler scene

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12378