[od6.2] Fixed rgb vs colorname mismatch
Summary:
When initializing a color input field with a random color, the color is picked from stddrawcols. These colors did not match colors used in the list of color descriptions. RGB(140,35,80) for instance is violet but was described as brown.
Changed each of the stddrawcols to match a color from the description list.
JIRA: OD-1944

Test Plan: Opened 'New WellTrack' a couple of times to verify color matches the name.

Differential Revision: https://phabricator.dgbes.com:8443/D12374