od6.2: Fixed double snapping in random line path generation
Summary: The values in the slower (smaller) dimension were snapped twice. For example, a value of 51.4 would get snapped to the nearest integer 51 first and then if you have a step of 2, it would again snap to 50. Now with a single snap it rightly ends up at 52.

Test Plan: tested the path of random line sent by Fernando Bordignon and it looks perfect now.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D12414