od6.2: have cmake test the actual usability of both svn and git during configuration
Summary: find_package is left to Quiet mode to avoid printing to many message. But we have test if both svn and git can be run since both are mandatory to checkout the externals.

Test Plan: replaced the value GIT_EXEC but a non-existing name to test that the configuration stops early enough.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D12433