[od6.2] Bugfix for zero width of log displays in surveys where XY is in feet
Summary:
In surveys where XY are in feet, logs always came up with a zero width.
Main problem was a wrong (too early) cast to int. Other problem is the conversion between feet and meter. The width is now stored as a float to minimize the conversion errors.

Test Plan: Log width is correctly saved in disp file and shows correct width at initial display.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12427